+++ src/libam_support/utils/version_utils.cc
@@ -59,6 +59,12 @@
 #include <ngs/Package.hpp>
 #endif
 
+#ifdef WITH_LIBDEFLATE
+#include <libdeflate.h>
+#endif
+
+#include <zlib.h>
+
 // In older versions it will be <fmt/core.h> with later versions <fmt/base.h>
 #include <fmt/format.h> // NOLINT
 
@@ -237,6 +243,16 @@ namespace {
 #endif
     }
 
+    void print_zlib_version() { std::cout << "zlib: " << zlibVersion() << std::endl; }
+
+    void print_libdeflate_version()
+    {
+#ifdef WITH_LIBDEFLATE
+        std::cout << "libdeflate: " << LIBDEFLATE_VERSION_STRING << std::endl;
+#else
+        std::cout << "libdeflate: not used" << std::endl;
+#endif
+    }
 } // namespace
 
 void print_version()
@@ -245,6 +261,8 @@ void print_version()
     std::cout << "ART_MODERN_LINK_LIBS: " << ART_MODERN_LINK_LIBS << std::endl;
     print_git_info();
     print_htslib_version();
+    print_zlib_version();
+    print_libdeflate_version();
     print_fmt_version();
     print_boost_version();
     print_onemkl_version();